Create ProgressCard component, move code that renders cards there, remove the same code from the Progress Logs page.
