+++ src/pages/ProgressLogs.css
@@ -22,6 +22,14 @@
   margin-bottom: 1rem;
 }
 
+.card__month {
+  cursor: pointer;
+}
+
+.card__year {
+  cursor: pointer;
+}
+
 .card__streak {
   padding: 0.5rem;
   margin: 0.5rem;
